requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
pyzabbix==1.3.1
asyncio==3.4.3
redis==5.0.1
pytz==2024.1 